requirements.txt
flask==3.0.3
psycopg[binary]>=3.2.0,<4
python-dotenv==1.0.1
werkzeug==3.0.3
gunicorn==22.0.0
